removed Duplications
