+++ plugman-core/src/main/java/core/com/rylinaux/plugman/commands/executables/RestartCommand.java
@@ -26,22 +26,15 @@ package core.com.rylinaux.plugman.commands.executables;
  * #L%
  */
 
-import core.com.rylinaux.plugman.commands.AbstractCommand;
 import core.com.rylinaux.plugman.commands.CommandSender;
-import core.com.rylinaux.plugman.plugins.Plugin;
 import core.com.rylinaux.plugman.services.ServiceRegistry;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
 
 /**
  * Command that restarts plugin(s).
  *
  * @author rylinaux
  */
-public class RestartCommand extends AbstractCommand {
+public class RestartCommand extends CascadingPluginCommand {
 
     /**
      * The name of the command.
@@ -77,144 +70,18 @@ public class RestartCommand extends AbstractCommand {
         super(sender, NAME, DESCRIPTION, PERMISSION, SUB_PERMISSIONS, USAGE, registry);
     }
 
-    /**
-     * Execute the command.
-     *
-     * @param sender the sender of the command
-     * @param label  the name of the command
-     * @param args   the arguments supplied
-     */
     @Override
-    public void execute(CommandSender sender, String label, String[] args) {
-        if (!validateArguments(label, args, 2)) return;
-
-        if (args[1].equalsIgnoreCase("all") || args[1].equalsIgnoreCase("*")) {
-            if (!hasPermission("all")) {
-                sendNoPermissionMessage();
-                return;
-            }
-            restartAllPlugins(sender, label);
-            return;
-        }
-
-        var target = getPluginManager().getPluginByName(args, 1);
-
-        if (!validatePlugin(label, target)) return;
-
-        restartPluginWithCommandBatch(sender, label, target);
+    protected String allSuccessMessage() {
+        return "restart.all";
     }
 
-    private void restartAllPlugins(CommandSender sender, String label) {
-        var plugins = getPluginManager().getPlugins().stream().filter(plugin ->
-                plugin != null && !getPluginManager().isIgnored(plugin)).toList();
-
-        var failedPlugins = new ArrayList<String>();
-
-        for (var plugin : plugins) {
-            var success = restartPluginWithCommandBatch(sender, label, plugin);
-
-            if (success) continue;
-            failedPlugins.add(plugin.getName());
-        }
-
-        if (failedPlugins.isEmpty()) {
-            sender.sendMessage("restart.all");
-            return;
-        }
-
-        sender.sendMessage("restart.all-failed", String.join(", ", failedPlugins));
+    @Override
+    protected String allFailedMessage() {
+        return "restart.all-failed";
     }
 
-    private boolean restartPluginWithCommandBatch(CommandSender sender, String label, Plugin target) {
-        getPluginManager().beginCommandUpdateBatch();
-        try {
-            return restartPlugin(sender, label, target);
-        } finally {
-            getPluginManager().endCommandUpdateBatch();
-        }
-    }
-
-    private boolean restartPlugin(CommandSender sender, String label, Plugin target) {
-        if (target == null) {
-            sendInvalidPluginMessage();
-            sendUsage(label);
-            return false;
-        }
-
-        var dependents = findEnabledDependents(target);
-        var unloadedDependents = new ArrayList<Plugin>();
-
-        for (var dependent : dependents) {
-            var result = getPluginManager().unload(dependent);
-            if (!result.success()) {
-                sender.sendMessage(result.messageId(), dependent.getName());
-                restartDependents(sender, unloadedDependents);
-                return false;
-            }
-            unloadedDependents.add(dependent);
-        }
-
-        var result = getPluginManager().unload(target);
-        if (!result.success()) {
-            sender.sendMessage(result.messageId(), target.getName());
-            restartDependents(sender, unloadedDependents);
-            return false;
-        }
-
-        result = getPluginManager().load(target);
-        if (!result.success()) {
-            sender.sendMessage(result.messageId(), result.messageArgs().length == 0 ? new Object[]{target.getName()} : result.messageArgs());
-            restartDependents(sender, unloadedDependents);
-            return false;
-        }
-
-        if (!restartDependents(sender, unloadedDependents)) return false;
-
-        sender.sendMessage("restart.restarted", target.getName());
-        return true;
-    }
-
-    private boolean restartDependents(CommandSender sender, List<Plugin> dependents) {
-        for (var i = dependents.size() - 1; i >= 0; i--) {
-            var dependent = dependents.get(i);
-
-            var result = getPluginManager().load(dependent);
-            if (result.success()) continue;
-
-            sender.sendMessage(result.messageId(), result.messageArgs().length == 0 ? new Object[]{dependent.getName()} : result.messageArgs());
-            return false;
-        }
-
-        return true;
-    }
-
-    private List<Plugin> findEnabledDependents(Plugin target) {
-        var dependents = new ArrayList<Plugin>();
-        var visited = new TreeSet<String>(String.CASE_INSENSITIVE_ORDER);
-        visited.add(target.getName());
-        collectEnabledDependents(target.getName(), dependents, visited);
-        return dependents;
-    }
-
-    private void collectEnabledDependents(String targetName, List<Plugin> dependents, Set<String> visited) {
-        for (var plugin : getPluginManager().getPlugins()) {
-            if (plugin == null || !plugin.isEnabled()) continue;
-            if (getPluginManager().isIgnored(plugin)) continue;
-            if (visited.contains(plugin.getName())) continue;
-            if (!dependsOn(plugin, targetName)) continue;
-
-            visited.add(plugin.getName());
-            collectEnabledDependents(plugin.getName(), dependents, visited);
-            dependents.add(plugin);
-        }
-    }
-
-    private boolean dependsOn(Plugin plugin, String targetName) {
-        return containsIgnoreCase(plugin.getDepend(), targetName)
-                || containsIgnoreCase(plugin.getSoftDepend(), targetName);
-    }
-
-    private boolean containsIgnoreCase(List<String> values, String expected) {
-        return values != null && values.stream().anyMatch(value -> value.equalsIgnoreCase(expected));
+    @Override
+    protected String pluginSuccessMessage() {
+        return "restart.restarted";
     }
 }
